+++ Arma3CfgFunctions/out/extension.js
@@ -1,4 +1,23 @@
 "use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
+    Object.defineProperty(o, "default", { enumerable: true, value: v });
+}) : function(o, v) {
+    o["default"] = v;
+});
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -8,11 +27,18 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
+};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.deactivate = exports.activate = void 0;
-const vscode = require("vscode");
-const fs = require("fs");
+const vscode = __importStar(require("vscode"));
+const fs = __importStar(require("fs"));
 const class_parser_1 = require("./class-parser");
+const axios_1 = __importDefault(require("axios"));
+const cheerio_1 = require("cheerio");
+const commands_json_1 = __importDefault(require("./Data/commands.json"));
+const functions_json_1 = __importDefault(require("./Data/functions.json"));
 //base defines
 let functionsLib = {};
 let config = vscode.workspace.getConfiguration('Arma3CfgFunctions');
@@ -25,37 +51,47 @@ function disposCompletionItems() {
     completionItems = [];
 }
 ;
-function updateMissionRoot() {
+let missionRoot = '';
+let descriptionPath = '';
+let workingFolderPath = '';
+function updateFolderPaths() {
     let workspaceFolders = vscode.workspace.workspaceFolders;
     if (workspaceFolders === undefined) {
-        return "";
+        return;
     }
     ;
-    let missionRoot = workspaceFolders[0].uri.fsPath;
+    missionRoot = workspaceFolders[0].uri.fsPath;
     if (config.has("MissionRoot")) {
         missionRoot = (missionRoot + "/" + config.get("MissionRoot")).split('\\').join('/');
     }
     ;
     console.debug("Mission root path: ", missionRoot);
-    return missionRoot;
+    if (config.has('DescriptionPath')) {
+        let path = config.get('DescriptionPath');
+        if (path != "") {
+            descriptionPath = path.split('\\').join('/');
+        }
+        else {
+            descriptionPath = '**/*description.ext';
+        }
+        ;
+    }
+    ;
+    console.debug("Description path: ", descriptionPath);
+    workingFolderPath = workspaceFolders[0].uri.fsPath.split('\\').join('/') + '/.vscode';
+    if (!fs.existsSync(workingFolderPath)) {
+        fs.mkdirSync(workingFolderPath);
+    }
+    ;
+    console.debug("Working folder path: ", workingFolderPath);
 }
 ;
 if (vscode.workspace.workspaceFolders !== undefined) {
-    updateMissionRoot();
+    updateFolderPaths();
 }
 ;
 function parseDescription(context) {
     return __awaiter(this, void 0, void 0, function* () {
-        //root description.ext path
-        let descriptionPath = '**/*description.ext';
-        if (config.has('DescriptionPath')) {
-            let path = config.get('DescriptionPath');
-            if (path != "") {
-                descriptionPath = path.split('\\').join('/');
-            }
-            ;
-        }
-        ;
         let description = yield vscode.workspace.findFiles(descriptionPath, "", 1);
         if (description.length == 0) {
             vscode.window.showWarningMessage("Arma3 CfgFunctions | Can't find description.ext, aborting!");
@@ -63,14 +99,6 @@ function parseDescription(context) {
         }
         ;
         console.info(`Description.ext path: ${description[0].fsPath}`);
-        //mission root
-        let missionRoot = updateMissionRoot();
-        //working folder
-        let workingFolderPath = missionRoot + '/.vscode';
-        if (!fs.existsSync(workingFolderPath)) {
-            fs.mkdirSync(workingFolderPath);
-        }
-        ;
         vscode.workspace.openTextDocument(description[0]).then((document) => __awaiter(this, void 0, void 0, function* () {
             if (document.isDirty) {
                 vscode.window.showInformationMessage(`Arma3 CfgFunctions | File unsaved, aborting. | Path: ${descriptionPath}`);
@@ -82,8 +110,8 @@ function parseDescription(context) {
             let brackets = 0;
             console.debug("Parsing flags declaired");
             //create temp file
-            fs.writeFileSync(workingFolderPath + "/cfgfunctions", "");
-            let fd = fs.openSync(workingFolderPath + "/cfgfunctions", "a+");
+            fs.writeFileSync(workingFolderPath + "/cfgfunctions.txt", "");
+            let fd = fs.openSync(workingFolderPath + "/cfgfunctions.txt", "a+");
             //write to temp file
             for (let index = 0; index < document.lineCount; index++) {
                 const element = document.lineAt(index);
@@ -120,8 +148,8 @@ function parseDescription(context) {
             ;
             fs.closeSync(fd);
             //parse to JSON with external lib and deleteTmpFile
-            let parsedjson = class_parser_1.parse(fs.readFileSync(workingFolderPath + "/cfgfunctions").toString());
-            fs.unlinkSync(workingFolderPath + "/cfgfunctions");
+            let parsedjson = class_parser_1.parse(fs.readFileSync(workingFolderPath + "/cfgfunctions.txt").toString());
+            fs.unlinkSync(workingFolderPath + "/cfgfunctions.txt");
             functionsLib = yield generateLibrary(parsedjson);
             console.log(functionsLib);
             //reload language additions (auto completion and header hovers)
@@ -192,7 +220,6 @@ function generateLibrary(cfgFunctionsJSON) {
             ;
         }
         ;
-        let missionRoot = updateMissionRoot();
         console.log("Generating function lib");
         let functionLib = {};
         //Default attributes
@@ -246,30 +273,6 @@ function generateLibrary(cfgFunctionsJSON) {
         ;
         return functionLib;
     });
-}
-;
-function PeekFile() {
-    //find what is selected
-    const editor = vscode.window.activeTextEditor;
-    let selectionStart = editor.selection.start;
-    if (!selectionStart) {
-        selectionStart = editor.selection.active;
-    }
-    ;
-    let wordRange = editor.document.getWordRangeAtPosition(selectionStart);
-    let selectedText = editor.document.getText(wordRange).toLowerCase();
-    //find function in library
-    let index = Object.getOwnPropertyNames(functionsLib).find((value) => value.toLowerCase() == selectedText);
-    if (index === undefined) {
-        vscode.window.showInformationMessage(`Arma3 CfgFunctions | Could not find function definition: ${selectedText}`);
-        return;
-    }
-    ;
-    //open file
-    let missionRoot = updateMissionRoot();
-    let functionPath = (missionRoot + '/' + functionsLib[index]['file']).split('/').join('\\');
-    let fncUri = vscode.Uri.file(functionPath);
-    vscode.window.showTextDocument(fncUri);
 }
 ;
 function getHeader(uri) {
@@ -378,9 +381,276 @@ function onSave(Document) {
     ;
 }
 ;
+function parseResponce(responce, classID) {
+    let $ = cheerio_1.load(responce.data);
+    //icon class
+    let icons = $('div.locality-icons');
+    let locIcons;
+    if (icons.length > 0) {
+        locIcons = icons[0]['children'];
+    }
+    ;
+    //header contents
+    let elements = $(classID)[0]['children'].filter(element => { return element.type == 'tag' && element.name == 'dl'; });
+    //parse text
+    let commandText = '';
+    function addToCommandText(text) { commandText += text; }
+    ;
+    function parseTypeDT(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) {
+                    addToCommandText(element.data);
+                }
+                ;
+            });
+        }
+    }
+    ;
+    function extractDataFromTag(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) {
+                    addToCommandText(element.data);
+                }
+                else {
+                    extractDataFromTag(element);
+                }
+                ;
+            });
+        }
+        ;
+    }
+    ;
+    function parseTypeDD(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) {
+                    addToCommandText(element.data);
+                }
+                else {
+                    switch (element.name) {
+                        case 'code':
+                            {
+                                addToCommandText('\n```sqf\n');
+                                extractDataFromTag(element);
+                                addToCommandText('\n```\n');
+                                break;
+                            }
+                            ;
+                        case 'a':
+                            {
+                                parseTypeA(element);
+                                break;
+                            }
+                            ;
+                        case 'b':
+                            {
+                                parseTypeB(element);
+                                break;
+                            }
+                            ;
+                        case 'tt':
+                            {
+                                parseTypeTT(element);
+                                break;
+                            }
+                            ;
+                    }
+                    ;
+                }
+                ;
+            });
+        }
+        ;
+    }
+    ;
+    function parseTypeTT(element) {
+        if (element.data) {
+            addToCommandText(element.data);
+        }
+        else {
+            if (element.children) {
+                element.children.forEach(element => { parseTypeTT(element); });
+            }
+            ;
+        }
+        ;
+    }
+    ;
+    function parseTypeB(element) {
+        element.children.forEach(element => {
+            if (element.data) {
+                addToCommandText(element.data);
+            }
+            else {
+                switch (element.type) {
+                    case 'tag':
+                        {
+                            extractDataFromTag(element);
+                            break;
+                        }
+                        ;
+                    default:
+                        {
+                            console.warn(`Unknown behaviour parseTypeB() -->`);
+                            console.warn(element);
+                            break;
+                        }
+                        ;
+                }
+                ;
+            }
+            ;
+        });
+    }
+    ;
+    function parseTypeA(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) {
+                    addToCommandText(element.data);
+                }
+                else {
+                    parseTypeA(element);
+                }
+                ;
+            });
+        }
+        ;
+    }
+    ;
+    function parseTypeDL(element) {
+        element.children.forEach(element => {
+            if (element.type == 'tag') {
+                switch (element.name) {
+                    case 'dt':
+                        {
+                            addToCommandText('\n\n### ');
+                            parseTypeDT(element);
+                            break;
+                        }
+                        ;
+                    case 'dd':
+                        {
+                            addToCommandText('\n');
+                            parseTypeDD(element);
+                            break;
+                        }
+                        ;
+                }
+            }
+        });
+    }
+    ;
+    if (locIcons) {
+        locIcons.forEach(icon => {
+            let title = icon['children'][0]['attribs']['title'];
+            if (title) {
+                addToCommandText(`[${title}]`);
+            }
+            ;
+        });
+    }
+    ;
+    elements.forEach(element => {
+        if (element.name == 'dl') {
+            parseTypeDL(element);
+            addToCommandText('\n');
+        }
+        else {
+            console.warn(`unknown tag name: ${element.name}`);
+        }
+        ;
+    });
+    commandText = commandText.substr(0, commandText.search('See also:')).trim();
+    return commandText;
+}
+;
+function generateCommands() {
+    return __awaiter(this, void 0, void 0, function* () {
+        const AxiosInstance = axios_1.default.create();
+        let output = {};
+        //get array of commands from input file
+        let fd = fs.openSync(`${workingFolderPath}/cfgFunctions/commands.txt`, "a+");
+        let A3Commands = fs.readFileSync(fd).toString().split(' ').join('_');
+        fs.closeSync(fd);
+        let commandsArray = A3Commands.split('\r\n');
+        if (commandsArray[commandsArray.length - 1] === '') {
+            commandsArray.pop();
+        }
+        ;
+        for (const element of commandsArray) {
+            console.log(element);
+            let commandUrl = `https://community.bistudio.com/wiki?title=${element}&printable=yes`;
+            let responce = yield AxiosInstance.get(commandUrl);
+            if (responce) {
+                let hover = parseResponce(responce, 'div._description.cmd');
+                output[element] = {
+                    Hover: hover,
+                    Url: `https://community.bistudio.com/wiki${element}`
+                };
+            }
+            ;
+        }
+        ;
+        console.log(output);
+        let commandsOutPath = `${workingFolderPath}/cfgFunctions/commands.json`;
+        if (fs.existsSync(commandsOutPath)) {
+            fs.unlinkSync(commandsOutPath);
+        }
+        fd = fs.openSync(commandsOutPath, "a+");
+        fs.writeFileSync(fd, JSON.stringify(output));
+        fs.closeSync(fd);
+    });
+}
+;
+function generateFunctions() {
+    return __awaiter(this, void 0, void 0, function* () {
+        const AxiosInstance = axios_1.default.create();
+        let output = {};
+        //get array of commands from input file
+        let fd = fs.openSync(`${workingFolderPath}/cfgFunctions/functions.txt`, "a+");
+        let A3Commands = fs.readFileSync(fd).toString().split(' ').join('_');
+        fs.closeSync(fd);
+        let commandsArray = A3Commands.split('\r\n');
+        if (commandsArray[commandsArray.length - 1] === '') {
+            commandsArray.pop();
+        }
+        ;
+        for (const element of commandsArray) {
+            if (!element.includes('_')) {
+                continue;
+            }
+            ; //Alfabettisation of list
+            console.log(element);
+            let commandUrl = `https://community.bistudio.com/wiki?title=${element}&printable=yes`;
+            let responce = yield AxiosInstance.get(commandUrl);
+            if (responce) {
+                let hover = parseResponce(responce, 'div._description.fnc');
+                output[element] = {
+                    Hover: hover,
+                    Url: `https://community.bistudio.com/wiki${element}`
+                };
+            }
+            ;
+        }
+        ;
+        console.log(output);
+        let commandsOutPath = `${workingFolderPath}/cfgFunctions/functions.json`;
+        if (fs.existsSync(commandsOutPath)) {
+            fs.unlinkSync(commandsOutPath);
+        }
+        fd = fs.openSync(commandsOutPath, "a+");
+        fs.writeFileSync(fd, JSON.stringify(output));
+        fs.closeSync(fd);
+    });
+}
+;
 function activate(context) {
     context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.recompile', () => parseDescription(context)));
-    context.subscriptions.push(vscode.commands.registerTextEditorCommand('a3cfgfunctions.peek', () => PeekFile()));
+    context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.generateCommands', () => generateCommands()));
+    context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.generateFunctions', () => generateFunctions()));
     if (vscode.workspace.workspaceFolders !== undefined) {
         parseDescription(context);
     }
@@ -394,8 +664,97 @@ function activate(context) {
     context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
         //update config and recompile with new settings
         config = vscode.workspace.getConfiguration('Arma3CfgFunctions');
+        updateFolderPaths();
         vscode.commands.executeCommand("a3cfgfunctions.recompile");
     }));
+    //commands wiki definitions
+    const commandsJsonKeys = Object.keys(commands_json_1.default);
+    const functionsJsonKeys = Object.keys(functions_json_1.default);
+    context.subscriptions.push(vscode.languages.registerDefinitionProvider({ language: 'sqf' }, {
+        provideDefinition(document, position) {
+            let word = document.getText(document.getWordRangeAtPosition(position));
+            //case sensetive checks (quick)
+            //engine commands
+            if (commands_json_1.default[word]) {
+                vscode.env.openExternal(commands_json_1.default[word].Url);
+                return undefined;
+            }
+            ;
+            //BIS functions
+            if (functions_json_1.default[word]) {
+                vscode.env.openExternal(functions_json_1.default[word].Url);
+                return undefined;
+            }
+            ;
+            //custom function
+            let key = Object.keys(functionsLib).find((Key) => { return Key.toLowerCase() == word.toLowerCase(); });
+            if (functionsLib[key]) {
+                return new vscode.Location(functionsLib[key].Uri, new vscode.Position(0, 0));
+            }
+            ;
+            if (config.get('caseInsensetive')) {
+                //engine commands
+                key = commandsJsonKeys.find((key) => { return key.toLowerCase() == word.toLowerCase(); });
+                if (commands_json_1.default[key]) {
+                    vscode.env.openExternal(commands_json_1.default[key].Url);
+                    return undefined;
+                }
+                ;
+                //BIS functions
+                key = functionsJsonKeys.find((key) => { return key.toLowerCase() == word.toLowerCase(); });
+                if (functions_json_1.default[key]) {
+                    vscode.env.openExternal(functions_json_1.default[key].Url);
+                    return undefined;
+                }
+                ;
+            }
+            return undefined;
+        }
+    }));
+    //commands wiki hovers (slows down hover loading drastically ~2500 hovers provided)
+    if (config.get('EnableCommandsHover')) {
+        for (const key in commands_json_1.default) {
+            if (Object.prototype.hasOwnProperty.call(commands_json_1.default, key)) {
+                const element = commands_json_1.default[key];
+                //hovers
+                context.subscriptions.push(vscode.languages.registerHoverProvider('sqf', {
+                    provideHover(document, position) {
+                        const range = document.getWordRangeAtPosition(position);
+                        const word = document.getText(range);
+                        if (word.toLowerCase() == key.toLowerCase()) {
+                            return new vscode.Hover(new vscode.MarkdownString(element.Hover));
+                        }
+                        ;
+                    }
+                }));
+            }
+            ;
+        }
+        ;
+    }
+    ;
+    //functions wiki hovers
+    if (config.get('EnableFunctionHover')) {
+        for (const key in functions_json_1.default) {
+            if (Object.prototype.hasOwnProperty.call(functions_json_1.default, key)) {
+                const element = functions_json_1.default[key];
+                //hovers
+                context.subscriptions.push(vscode.languages.registerHoverProvider('sqf', {
+                    provideHover(document, position) {
+                        const range = document.getWordRangeAtPosition(position);
+                        const word = document.getText(range);
+                        if (word.toLowerCase() == key.toLowerCase()) {
+                            return new vscode.Hover(new vscode.MarkdownString(element.Hover));
+                        }
+                        ;
+                    }
+                }));
+            }
+            ;
+        }
+        ;
+    }
+    ;
 }
 exports.activate = activate;
 function deactivate() {