+++ Arma3CfgFunctions/src/extension.ts
@@ -1,9 +1,12 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 import {parse} from './class-parser';
+import axios from "axios";
+import { load } from "cheerio";
+import commandsJson from "./Data/commands.json";
+import functionsJson from "./Data/functions.json";
 
 //base defines
-
 let functionsLib = {}
 let config = vscode.workspace.getConfiguration('Arma3CfgFunctions');
 let completionItems: vscode.Disposable[] = [];
@@ -16,41 +19,43 @@ function disposCompletionItems() {
     completionItems = [];
 };
 
-function updateMissionRoot() {
+let missionRoot = '';
+let descriptionPath = '';
+let workingFolderPath = '';
+function updateFolderPaths() {
     let workspaceFolders = vscode.workspace.workspaceFolders;
-    if (workspaceFolders === undefined) { return "" };
-    let missionRoot = workspaceFolders[0].uri.fsPath;
+    if (workspaceFolders === undefined) { return };
+    missionRoot = workspaceFolders[0].uri.fsPath;
     if (config.has("MissionRoot")) {
         missionRoot = (missionRoot + "/" + config.get("MissionRoot")).split('\\').join('/');
     };
     console.debug("Mission root path: ", missionRoot);
-    return missionRoot;
+
+
+    if (config.has('DescriptionPath')) {
+        let path = <string>config.get('DescriptionPath');
+        if (path != "") {
+            descriptionPath = path.split('\\').join('/');
+        } else {
+            descriptionPath = '**/*description.ext';
+        };
+    };
+    console.debug("Description path: ", descriptionPath);
+
+    workingFolderPath = workspaceFolders[0].uri.fsPath.split('\\').join('/') + '/.vscode';
+    if (!fs.existsSync(workingFolderPath)) {
+        fs.mkdirSync(workingFolderPath);
+    };
+    console.debug("Working folder path: ", workingFolderPath);
 };
-if (vscode.workspace.workspaceFolders !== undefined) { updateMissionRoot() };
+if (vscode.workspace.workspaceFolders !== undefined) { updateFolderPaths() };
 
 
 async function parseDescription(context: vscode.ExtensionContext) {
 
-    //root description.ext path
-    let descriptionPath = '**/*description.ext';
-    if (config.has('DescriptionPath')) {
-        let path = <string> config.get('DescriptionPath');
-        if (path != "") {
-            descriptionPath = path.split('\\').join('/');
-        };
-    };
     let description = await vscode.workspace.findFiles(descriptionPath, "", 1);
     if (description.length == 0) { vscode.window.showWarningMessage("Arma3 CfgFunctions | Can't find description.ext, aborting!"); return };
     console.info(`Description.ext path: ${description[0].fsPath}`);
-
-    //mission root
-    let missionRoot = updateMissionRoot();
-
-    //working folder
-    let workingFolderPath = missionRoot + '/.vscode';
-    if (!fs.existsSync(workingFolderPath)) {
-        fs.mkdirSync(workingFolderPath);
-    };
 
     vscode.workspace.openTextDocument(description[0]).then(async (document) => {
         if (document.isDirty) { vscode.window.showInformationMessage(`Arma3 CfgFunctions | File unsaved, aborting. | Path: ${descriptionPath}`);};
@@ -61,8 +66,8 @@ async function parseDescription(context: vscode.ExtensionContext) {
         console.debug("Parsing flags declaired");
 
         //create temp file
-        fs.writeFileSync(workingFolderPath+"/cfgfunctions", "");
-        let fd = fs.openSync(workingFolderPath+"/cfgfunctions", "a+");
+        fs.writeFileSync(workingFolderPath + "/cfgfunctions.txt", "");
+        let fd = fs.openSync(workingFolderPath + "/cfgfunctions.txt", "a+");
 
         //write to temp file
         for (let index = 0; index < document.lineCount; index++) {
@@ -88,8 +93,8 @@ async function parseDescription(context: vscode.ExtensionContext) {
         fs.closeSync(fd);
 
         //parse to JSON with external lib and deleteTmpFile
-        let parsedjson = parse(fs.readFileSync(workingFolderPath+"/cfgfunctions").toString());
-        fs.unlinkSync(workingFolderPath+"/cfgfunctions");
+        let parsedjson = parse(fs.readFileSync(workingFolderPath + "/cfgfunctions.txt").toString());
+        fs.unlinkSync(workingFolderPath + "/cfgfunctions.txt");
 
         functionsLib = await generateLibrary(parsedjson);
         console.log(functionsLib);
@@ -152,8 +157,6 @@ async function generateLibrary(cfgFunctionsJSON:JSON) {
         };
     };
 
-    let missionRoot = updateMissionRoot();
-
     console.log("Generating function lib");
     let functionLib = {};
 
@@ -212,27 +215,6 @@ async function generateLibrary(cfgFunctionsJSON:JSON) {
         }
     };
     return functionLib;
-};
-
-function PeekFile() {
-    //find what is selected
-    const editor = vscode.window.activeTextEditor;
-    let selectionStart = editor.selection.start;
-    if (!selectionStart) {
-        selectionStart = editor.selection.active;
-    };
-    let wordRange = editor.document.getWordRangeAtPosition(selectionStart);
-    let selectedText = editor.document.getText(wordRange).toLowerCase();
-
-    //find function in library
-    let index = Object.getOwnPropertyNames(functionsLib).find((value: string) => value.toLowerCase() == selectedText);
-    if (index === undefined) { vscode.window.showInformationMessage(`Arma3 CfgFunctions | Could not find function definition: ${selectedText}`); return };
-
-    //open file
-    let missionRoot = updateMissionRoot();
-    let functionPath = (missionRoot + '/' + functionsLib[index]['file']).split('/').join('\\');
-    let fncUri = vscode.Uri.file(functionPath);
-    vscode.window.showTextDocument(fncUri);
 };
 
 async function getHeader(uri: vscode.Uri) {
@@ -321,9 +303,190 @@ function onSave(Document: vscode.TextDocument) {
     };
 };
 
+function parseResponce(responce, classID) {
+    let $ = load(responce.data);
+
+    //icon class
+    let icons = $('div.locality-icons');
+    let locIcons: any;
+    if (icons.length > 0) { locIcons = icons[0]['children'] };
+
+    //header contents
+    let elements = $(classID)[0]['children'].filter(element => { return element.type == 'tag' && element.name == 'dl' });
+
+    //parse text
+    let commandText = '';
+    function addToCommandText(text: string) { commandText += text };
+
+    function parseTypeDT(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) { addToCommandText(element.data) };
+            });
+        }
+    };
+
+    function extractDataFromTag(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) { addToCommandText(element.data) } else {
+                    extractDataFromTag(element);
+                };
+            });
+        };
+    };
+
+    function parseTypeDD(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) { addToCommandText(element.data) } else {
+                    switch (element.name) {
+                        case 'code': { addToCommandText('\n```sqf\n'); extractDataFromTag(element); addToCommandText('\n```\n'); break };
+                        case 'a': { parseTypeA(element); break };
+                        case 'b': { parseTypeB(element); break };
+                        case 'tt': { parseTypeTT(element); break };
+                    };
+                };
+            });
+        };
+    };
+
+    function parseTypeTT(element) {
+        if (element.data) { addToCommandText(element.data); } else {
+            if (element.children) { element.children.forEach(element => { parseTypeTT(element) }) };
+        };
+    };
+
+    function parseTypeB(element) {
+        element.children.forEach(element => {
+            if (element.data) {
+                addToCommandText(element.data);
+            } else {
+                switch (element.type) {
+                    case 'tag': {
+                        extractDataFromTag(element);
+                        break
+                    };
+                    default: {
+                        console.warn(`Unknown behaviour parseTypeB() -->`);
+                        console.warn(element);
+                        break
+                    };
+                };
+            };
+        });
+    };
+
+    function parseTypeA(element) {
+        if (element.children) {
+            element.children.forEach(element => {
+                if (element.data) { addToCommandText(element.data) } else {
+                    parseTypeA(element);
+                };
+            });
+        };
+    };
+
+    function parseTypeDL(element) {
+        element.children.forEach(element => {
+            if (element.type == 'tag') {
+                switch (element.name) {
+                    case 'dt': { addToCommandText('\n\n### '); parseTypeDT(element); break };
+                    case 'dd': { addToCommandText('\n'); parseTypeDD(element); break };
+                }
+            }
+        });
+    };
+
+    if (locIcons) {
+        locIcons.forEach(icon => {
+            let title = icon['children'][0]['attribs']['title'];
+            if (title) {
+                addToCommandText(`[${title}]`);
+            };
+        });
+    };
+
+    elements.forEach(element => {
+        if (element.name == 'dl') { parseTypeDL(element); addToCommandText('\n') } else {
+            console.warn(`unknown tag name: ${element.name}`);
+        };
+    });
+
+    commandText = commandText.substr(0, commandText.search('See also:')).trim();
+    return commandText;
+};
+
+async function generateCommands() {
+    const AxiosInstance = axios.create();
+    let output = {};
+
+    //get array of commands from input file
+    let fd = fs.openSync(`${workingFolderPath}/cfgFunctions/commands.txt`, "a+");
+    let A3Commands = fs.readFileSync(fd).toString().split(' ').join('_');
+    fs.closeSync(fd);
+    let commandsArray = A3Commands.split('\r\n');
+    if (commandsArray[commandsArray.length - 1] === '') { commandsArray.pop() };
+
+    for (const element of commandsArray) {
+        console.log(element);
+        let commandUrl = `https://community.bistudio.com/wiki?title=${element}&printable=yes`;
+        let responce = await AxiosInstance.get(commandUrl);
+        if (responce) {
+            let hover = parseResponce(responce, 'div._description.cmd');
+            output[element] = {
+                Hover: hover,
+                Url: `https://community.bistudio.com/wiki${element}`
+            };
+        };
+    };
+
+    console.log(output);
+    let commandsOutPath = `${workingFolderPath}/cfgFunctions/commands.json`;
+    if (fs.existsSync(commandsOutPath)) { fs.unlinkSync(commandsOutPath) }
+    fd = fs.openSync(commandsOutPath, "a+");
+    fs.writeFileSync(fd, JSON.stringify(output));
+    fs.closeSync(fd);
+
+};
+
+async function generateFunctions() {
+    const AxiosInstance = axios.create();
+    let output = {};
+    //get array of commands from input file
+    let fd = fs.openSync(`${workingFolderPath}/cfgFunctions/functions.txt`, "a+");
+    let A3Commands = fs.readFileSync(fd).toString().split(' ').join('_');
+    fs.closeSync(fd);
+    let commandsArray = A3Commands.split('\r\n');
+    if (commandsArray[commandsArray.length - 1] === '') { commandsArray.pop() };
+
+    for (const element of commandsArray) {
+        if (!element.includes('_')) { continue }; //Alfabettisation of list
+        console.log(element);
+        let commandUrl = `https://community.bistudio.com/wiki?title=${element}&printable=yes`;
+        let responce = await AxiosInstance.get(commandUrl);
+        if (responce) {
+            let hover = parseResponce(responce, 'div._description.fnc');
+            output[element] = {
+                Hover: hover,
+                Url: `https://community.bistudio.com/wiki${element}`
+            };
+        };
+    };
+
+    console.log(output);
+    let commandsOutPath = `${workingFolderPath}/cfgFunctions/functions.json`;
+    if (fs.existsSync(commandsOutPath)) { fs.unlinkSync(commandsOutPath) }
+    fd = fs.openSync(commandsOutPath, "a+");
+    fs.writeFileSync(fd, JSON.stringify(output));
+    fs.closeSync(fd);
+
+};
+
 export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.recompile', () => parseDescription(context)));
-    context.subscriptions.push(vscode.commands.registerTextEditorCommand('a3cfgfunctions.peek', () => PeekFile()));
+    context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.generateCommands', () => generateCommands()));
+    context.subscriptions.push(vscode.commands.registerCommand('a3cfgfunctions.generateFunctions', () => generateFunctions()));
     if (vscode.workspace.workspaceFolders !== undefined) { parseDescription(context) };
 
     //events
@@ -337,8 +500,98 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(() => {
         //update config and recompile with new settings
         config = vscode.workspace.getConfiguration('Arma3CfgFunctions');
+        updateFolderPaths()
         vscode.commands.executeCommand("a3cfgfunctions.recompile");
     }))
+
+    //commands wiki definitions
+    const commandsJsonKeys = Object.keys(commandsJson);
+    const functionsJsonKeys = Object.keys(functionsJson);
+    context.subscriptions.push(vscode.languages.registerDefinitionProvider({ language: 'sqf' }, {
+        provideDefinition(document, position) {
+            let word = document.getText(document.getWordRangeAtPosition(position));
+
+            //case sensetive checks (quick)
+            //engine commands
+            if (commandsJson[word]) {
+                vscode.env.openExternal(commandsJson[word].Url);
+                return undefined
+            };
+            //BIS functions
+            if (functionsJson[word]) {
+                vscode.env.openExternal(functionsJson[word].Url);
+                return undefined
+            };
+
+            //custom function
+            let key = Object.keys(functionsLib).find((Key) => { return Key.toLowerCase() == word.toLowerCase() });
+            if (functionsLib[key]) {
+                return new vscode.Location(functionsLib[key].Uri, new vscode.Position(0, 0));
+            };
+
+            if (config.get('caseInsensetive')) {
+                //engine commands
+                key = commandsJsonKeys.find((key) => { return key.toLowerCase() == word.toLowerCase() });
+                if (commandsJson[key]) {
+                    vscode.env.openExternal(commandsJson[key].Url);
+                    return undefined
+                };
+
+                //BIS functions
+                key = functionsJsonKeys.find((key) => { return key.toLowerCase() == word.toLowerCase() });
+                if (functionsJson[key]) {
+                    vscode.env.openExternal(functionsJson[key].Url);
+                    return undefined
+                };
+            }
+
+            return undefined
+        }
+    }));
+
+    //commands wiki hovers (slows down hover loading drastically ~2500 hovers provided)
+    if (config.get('EnableCommandsHover')) {
+        for (const key in commandsJson) {
+            if (Object.prototype.hasOwnProperty.call(commandsJson, key)) {
+                const element = commandsJson[key];
+                //hovers
+                context.subscriptions.push(vscode.languages.registerHoverProvider('sqf', {
+                    provideHover(document, position) {
+                        const range = document.getWordRangeAtPosition(position);
+                        const word = document.getText(range);
+                        if (word.toLowerCase() == key.toLowerCase()) {
+
+                            return new vscode.Hover(
+                                new vscode.MarkdownString(element.Hover)
+                            );
+                        };
+                    }
+                }));
+            };
+        };
+    };
+
+    //functions wiki hovers
+    if (config.get('EnableFunctionHover')) {
+        for (const key in functionsJson) {
+            if (Object.prototype.hasOwnProperty.call(functionsJson, key)) {
+                const element = functionsJson[key];
+                //hovers
+                context.subscriptions.push(vscode.languages.registerHoverProvider('sqf', {
+                    provideHover(document, position) {
+                        const range = document.getWordRangeAtPosition(position);
+                        const word = document.getText(range);
+                        if (word.toLowerCase() == key.toLowerCase()) {
+
+                            return new vscode.Hover(
+                                new vscode.MarkdownString(element.Hover)
+                            );
+                        };
+                    }
+                }));
+            };
+        };
+    };
 }
 
 export function deactivate() {